rename files

+++ migrations/001_init.up.sql
@@ -1,14 +1,14 @@
 CREATE TABLE t_account (
-       account_id INTEGER NOT NULL PRIMARY KEY,
+       account_id BIGINT NOT NULL PRIMARY KEY,
        balance VARCHAR(255)
 );
 
 CREATE TABLE t_transaction (
      uuid UUID DEFAULT gen_random_uuid() NOT NULL,
-     source_account_id INTEGER NOT NULL,
+     source_account_id BIGINT NOT NULL,
      source_init VARCHAR(255) NOT NULL,
      source_after VARCHAR(255) NOT NULL,
-     destination_account_id INTEGER NOT NULL,
+     destination_account_id BIGINT NOT NULL,
      destination_init VARCHAR(255) NOT NULL,
      destination_after VARCHAR(255) NOT NULL,
      amount VARCHAR(255),
